UPDATE 3
 UPDATE 3

The big one! So to begin this I first converted the previous files, Bot_Build.py, ItemInfo.py, LoginCredentials.py, and URLCreation.py to a single 'mega' file BotBuild.py. I added the PriceCheck() function which implements BeautifulSoup webscraping functions to grab and format the specified items price, and checks that the price is between a minimum price parameter and a maximum price parameter. Next I rebuilt ProductCheck() function, reducing the number of lines from 21 to 14 while implementing PriceCheck() function as a parameter for the item to be added to cart. Finally, I structured a dictionary that holds all the item and user information for the bot to use in both sign in and ProductCheck().